Création de la methode tourelle agit et variable temps

+++ VirusMinator/src/application/modele/virus/VirusBasirus.java
@@ -2,7 +2,7 @@ package application.modele.virus;
 
 public class VirusBasirus extends Virus {
 
-	public VirusBasirus(int vie, int atq, double vitesse, String nom, int x, int y) {
+	public VirusBasirus(double vie, int atq, double vitesse, String nom, int x, int y) {
 		super(vie, atq, vitesse, nom, x, y);
 	}
