+++ VirusMinator/src/application/modele/virus/VirusDivirus.java
@@ -2,7 +2,7 @@ package application.modele.virus;
 
 public class VirusDivirus extends Virus {
 
-	public VirusDivirus(int vie, int atq, double vitesse, String nom, int x, int y) {
+	public VirusDivirus(double vie, int atq, double vitesse, String nom, int x, int y) {
 		super(vie, atq, vitesse, nom, x, y);
 	}
