+++ VirusMinator/src/application/modele/virus/VirusVhealrus.java
@@ -2,7 +2,7 @@ package application.modele.virus;
 
 public class VirusVhealrus extends Virus {
 
-	public VirusVhealrus(int vie, int atq, double vitesse, String nom, int x, int y) {
+	public VirusVhealrus(double vie, int atq, double vitesse, String nom, int x, int y) {
 		super(vie, atq, vitesse, nom, x, y);
 	}
