+++ VirusMinator/src/application/modele/virus/VirusViboomrus.java
@@ -2,7 +2,7 @@ package application.modele.virus;
 
 public class VirusViboomrus extends Virus {
 
-	public VirusViboomrus(int vie, int atq, double vitesse, String nom, int x, int y) {
+	public VirusViboomrus(double vie, int atq, double vitesse, String nom, int x, int y) {
 		super(vie, atq, vitesse, nom, x, y);
 	}
