+++ VirusMinator/src/application/modele/virus/VirusViterus.java
@@ -2,7 +2,7 @@ package application.modele.virus;
 
 public class VirusViterus extends Virus {
 
-	public VirusViterus(int vie, int atq, double vitesse, String nom, int x, int y) {
+	public VirusViterus(double vie, int atq, double vitesse, String nom, int x, int y) {
 		super(vie, atq, vitesse, nom, x, y);
 	}
